+++ lib/swagger-helpers.js
@@ -1,7 +1,7 @@
 "use strict";
 
 // Dependencies.
-const recursive = require("../utils/recursive");
+const RecursiveIterator = require("../utils/recursive");
 
 /**
  * Checks if tag is already contained withing target.
@@ -240,14 +240,15 @@ function addDataToSwaggerObject(swaggerObject, data) {
  * @param {Array} problems - aggregate list of found problems
  */
 function seekWrong(list, wrongSet, problems) {
-	recursive.recursiveIterate(list, (item) => {
+	let iterator = new RecursiveIterator(list, 0, false);
+	for (let item = iterator.next(); !item.done; item = iterator.next()) {
 		let isDirectChildOfProperties =
-			item.path[item.path.length - 2] === "properties";
+			item.value.path[item.value.path.length - 2] === "properties";
 
-		if (wrongSet.indexOf(item.key) > 0 && !isDirectChildOfProperties) {
-			problems.push(item.key);
+		if (wrongSet.indexOf(item.value.key) > 0 && !isDirectChildOfProperties) {
+			problems.push(item.value.key);
 		}
-	});
+	}
 }
 
 /**
