+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "talec-swagger",
-	"version": "0.0.5",
+	"version": "0.0.6",
 	"description": "Generates swagger doc & ui based on express existing routes.",
 	"main": "index.js",
 	"scripts": {
@@ -21,7 +21,8 @@
 	"license": "MIT",
 	"dependencies": {
 		"comment-parser": "^1.4.1",
-		"glob": "^7.2.0",
+		"doctrine-file": "^2.0.1",
+		"glob": "^10.3.16",
 		"swagger-parser": "^10.0.3",
 		"swagger-ui-express": "^5.0.0"
 	}
