--- test/swagger.json
@@ -1,21 +1,0 @@
-{
-  "info": {
-    "description": "This is swagger is a test to evolve skarn-swagger.",
-    "title": "Swagger Skarn test API",
-    "version": "1.0.0"
-  },
-  "servers": [
-    {
-      "url": "http://localhost:3000/",
-      "description": "localhost"
-    }
-  ],
-  "openapi": "3.0.0",
-  "paths": {},
-  "components": {},
-  "schemas": {},
-  "responses": {},
-  "parameters": {},
-  "securityDefinitions": {},
-  "tags": []
-}
